% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterthanfactoring.R
\name{betterthanfactoring}
\alias{betterthanfactoring}
\title{Factors specified vector of variables in a data frame.}
\usage{
betterthanfactoring(data, vectornames)
}
\arguments{
\item{data}{is the data frame the variables you want to factor are contained in.}

\item{vectornames}{is a vector containing the specific variable names you want to factor as strings.}
}
\value{
The updated data frame containing the factored variables.
}
\description{
Factors specified vector of variables in a data frame.
}
\examples{
neardeath <- readr::read_csv("https://raw.githubusercontent.com/rfordatascience/tidytuesday/refs/heads/main/data/2026/2026-07-21/nde_experiences.csv")
class(neardeath$gender)
test <- betterthanfactoring(neardeath, c("gender", "country"))
class(test$gender)
}
