% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_KCundercover.R
\name{theme_KCundercover}
\alias{theme_KCundercover}
\title{Theme function that can be applied to customize ggplots.}
\usage{
theme_KCundercover(gridline_x = TRUE, gridline_y = TRUE)
}
\arguments{
\item{gridline_x}{specifies if you would like x axis gridlines on your ggplot.}

\item{gridline_y}{specifies if you would like y axis gridlines on your ggplot.}
}
\value{
The ggplot created using the ggplot function customized to the KCundercover theme.
}
\description{
Theme function that can be applied to customize ggplots.
}
